personal-calls: move Container component into common/
We'll want to reuse it in `super-app`.
